# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "News Aggregator application API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get articles list with filters",
                "description": "Retrieve articles with optional filters and pagination",
                "operationId": "7e370449b00151b6422003b99ee09b88",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search in title, description and content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "technology"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "business",
                                "entertainment",
                                "general",
                                "health",
                                "science",
                                "sports",
                                "technology"
                            ],
                            "example": "technology"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter by news source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "newsapi"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by published date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-03-15"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Latest Tech News"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Description of the article"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Full content of the article"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://example.com/article"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "technology"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "newsapi"
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-03-15T12:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/{articleId}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get specific article",
                "description": "Retrieve a specific article by its ID",
                "operationId": "1f75031f2efdf1815aa8744011fefb76",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Latest Tech News"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Description of the article"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Full content of the article"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://example.com/article"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "technology"
                                        },
                                        "source": {
                                            "type": "string",
                                            "example": "newsapi"
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-03-15T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Article]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": {
        "1": {
            "sanctum": []
        }
    }
}